separated init tests, added a map test
